Make detection of scif.h work on SLES 11.x and 12.x distros.
The location of /lib/modules/`uname –r`/source differs on RHEL and SLES
distros, but older MPSS 3.x assumes they are the same and installs in the
wrong location. This previously worked because the internal scif.h was always
used, but this breaks when scif.h changes in MPSS 4.x.

Signed-off-by: Patrick McCormick patrick.m.mccormick@intel.com