Fixed scif compilation environment
Include <scif.h> instead of <modules/scif.h> and adjust the makefiles accordingly. This
has to be done because including the parent directory also ropes in the host kernel’s
include directories, which breaks the build of core on some configurations.

Check for the IBP and SCIF config keys when adding scif.h paths to CFLAGS, before
only the SCIF key was checked even though both depend on scif and can be compiled
separately.

Signed-off-by: Patrick McCormick patrick.m.mccormick@intel.com